various comment additions

+++ de.gebit.integrity.dsl/src/de/gebit/integrity/operations/standard/operands/DivisionNode.java
@@ -6,20 +6,26 @@ package de.gebit.integrity.operations.standard.operands;
 import java.math.BigDecimal;
 
 /**
+ * Divides two numbers. Uses double values as input since BigDecimals cannot always be divided, for example in case of
+ * periodic results.
  * 
- * 
- * @author Slartibartfast
+ * @author Rene Schneider
  * 
  */
-public class DivisionNode extends OperatorNode<BigDecimal, BigDecimal> {
+public class DivisionNode extends OperatorNode<Double, Double> {
 
+	/**
+	 * Creates a new instance.
+	 * 
+	 * @param aLeftOperand
+	 * @param aRightOperand
+	 */
 	public DivisionNode(Object aLeftOperand, Object aRightOperand) {
 		super(aLeftOperand, aRightOperand);
 	}
 
 	@Override
-	protected Object evaluateInternal(BigDecimal aLeftOperand, BigDecimal aRightOperand) {
-		// Conversion to double values is necessary because BigDecimal division fails hard in case of periodic results.
-		return new BigDecimal(aLeftOperand.doubleValue() / aRightOperand.doubleValue());
+	protected Object evaluateInternal(Double aLeftOperand, Double aRightOperand) {
+		return new BigDecimal(aLeftOperand / aRightOperand);
 	}
 }
